integrating with original method
